[AV-65208] Datasources for some resources (#57)
* Add datasources, rename datasources

* Generate docs

+++ internal/api/bucket/stats.go
@@ -2,8 +2,8 @@ package bucket
 
 // Stats are the bucket related statistics that are sent by the Capella V4 Public API for any existing bucket.
 type Stats struct {
-	ItemCount       int `json:"itemCount"`
-	OpsPerSecond    int `json:"opsPerSecond"`
-	DiskUsedInMib   int `json:"diskUsedInMib"`
-	MemoryUsedInMib int `json:"memoryUsedInMib"`
+	ItemCount       int64 `json:"itemCount"`
+	OpsPerSecond    int64 `json:"opsPerSecond"`
+	DiskUsedInMib   int64 `json:"diskUsedInMib"`
+	MemoryUsedInMib int64 `json:"memoryUsedInMib"`
 }
